add new folder and move the GetAllDogsQuery and GetAllDogsQueryHandler in.
