+++ Application/Queries/Dogs/GetAllDogs/GetAllDogsQuery.cs
@@ -1,7 +1,7 @@
 ﻿using MediatR;
 using Domain.Modulea;
 
-namespace Application.Queries.Dogs
+namespace Application.Queries.Dogs.GetAllDogs
 {
     public class GetAllDogsQuery : IRequest<List<Dog>>
     {
